requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
cryptography
python-binance